requirements.txt
fastapi
uvicorn
sqlmodel
python-dotenv
pydantic[email]
pyjwt
bcrypt==4.0.1
passlib[bcrypt]==1.7.4